fix(tests): Let the server run after approval tests
